+++ types/events.d.ts
@@ -1,11 +1,14 @@
 export declare enum UIEvent {
     EXPORT = "export",
     CLOSE = "close",
+    HISTORY_UNDO = "historyUndo",
+    HISTORY_REDO = "historyRedo",
     HISTORY_CHANGE = "historyChange",
     EDITOR_READY = "editorReady",
     ERROR_IMAGE_LOADING = "errorImageLoading",
     IMAGE_LOAD = "imageLoad",
     TOOL_ENTER = "toolEnter",
     STICKER_UPLOAD_ADD = "stickerUploadAdd",
-    ERROR_WEBGL_CONTEXT_LOST = "errorWebGLContextLost"
+    ERROR_WEBGL_CONTEXT_LOST = "errorWebGLContextLost",
+    CROP_DIMENSIONS_CHANGE = "cropDimensionsChange"
 }